Test to remove dynamic param

+++ src/public/Install-NerdFont.ps1
@@ -25,47 +25,22 @@ function Install-NerdFont {
         Installs all Nerd Fonts to the current user.
     #>
     [CmdletBinding(
-        DefaultParameterSetName = 'Name'
+        DefaultParameterSetName = 'Name',
+        SupportsShouldProcess
     )]
     [Alias('Install-NerdFonts')]
     param(
-        [Parameter(
-            Mandatory,
-            Position = 0,
-            ParameterSetName = 'All'
-        )]
+        [Parameter(ParameterSetName = 'All', Mandatory)]
         [switch] $All,
 
-        [Parameter(
-            Position = 1,
-            ParameterSetName = '__AllParameterSets'
-        )]
+        [Parameter(ParameterSetName = 'Name', Mandatory)]
+        [ValidateSet({ Get-NerdFonts | Select-Object -ExpandProperty Name })]
+        [string[]] $Name,
+
+        [Parameter()]
         [ValidateSet('CurrentUser', 'AllUsers')]
         [string] $Scope = 'CurrentUser'
     )
-
-    DynamicParam {
-        $runtimeDefinedParameterDictionary = New-Object System.Management.Automation.RuntimeDefinedParameterDictionary
-        $attributeCollection = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
-
-        $parameterName = 'Name'
-        $parameterAttribute = New-Object System.Management.Automation.ParameterAttribute
-        $parameterAttribute.Mandatory = $true
-        $parameterAttribute.ParameterSetName = 'Name'
-        $parameterAttribute.Position = 0
-        $parameterAttribute.HelpMessage = 'Name of the font to uninstall.'
-        $parameterAttribute.ValueFromPipeline = $true
-        $parameterAttribute.ValueFromPipelineByPropertyName = $true
-        $attributeCollection.Add($parameterAttribute)
-
-        $parameterValidateSet = (Get-NerdFonts).Name
-        $validateSetAttribute = New-Object System.Management.Automation.ValidateSetAttribute($parameterValidateSet)
-        $attributeCollection.Add($validateSetAttribute)
-
-        $runtimeDefinedParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($parameterName, [string[]], $attributeCollection)
-        $runtimeDefinedParameterDictionary.Add($parameterName, $runtimeDefinedParameter)
-        return $runtimeDefinedParameterDictionary
-    }
 
     begin {
         if ($Scope -eq 'AllUsers' -and -not (IsAdmin)) {
@@ -100,16 +75,22 @@ Please run the command again with elevated rights (Run as Administrator) or prov
             Write-Verbose "[$FontName] - Downloading to [$downloadPath]"
             $storedProgressPreference = $ProgressPreference
             $ProgressPreference = 'SilentlyContinue' # Suppress progress bar
-            Invoke-WebRequest -Uri $URL -OutFile $downloadPath -Verbose:$false
+            if ($PSCmdlet.ShouldProcess($FontName, "Download $FontName")) {
+                Invoke-WebRequest -Uri $URL -OutFile $downloadPath -Verbose:$false
+            }
             $ProgressPreference = $storedProgressPreference
 
             Write-Verbose "[$FontName] - Unpack to [$extractPath]"
-            Expand-Archive -Path $downloadPath -DestinationPath $extractPath -Force
-            Remove-Item -Path $downloadPath -Force
+            if ($PSCmdlet.ShouldProcess($FontName, 'Extract archive')) {
+                Expand-Archive -Path $downloadPath -DestinationPath $extractPath -Force
+                Remove-Item -Path $downloadPath -Force
+            }
 
             Write-Verbose "[$FontName] - Install to [$Scope]"
-            Install-Font -Path $extractPath -Scope $Scope
-            Remove-Item -Path $extractPath -Force -Recurse
+            if ($PSCmdlet.ShouldProcess($FontName, 'Install font')) {
+                Install-Font -Path $extractPath -Scope $Scope
+                Remove-Item -Path $extractPath -Force -Recurse
+            }
         }
     }
